내부 private 구현을 숨기는 구현 4
